+++ src/Cooked/MockChain/GenerateTx/ReferenceInputs.hs
@@ -13,9 +13,9 @@ import Polysemy.Error
 
 -- | Takes a 'TxSkel' and generates the associated 'Cardano.TxInsReference' from
 -- its content. These reference inputs can be found in two places, either in
--- direct reference inputs 'txSkelInsReference' or scattered in the various
+-- direct reference inputs 'txSkelReferenceInputs' or scattered in the various
 -- redeemers of the transaction, which can be gathered with
--- 'txSkelInsReferenceInRedeemers'.
+-- 'txSkelReferenceInputsInRedeemers'.
 toInsReference ::
   (Members '[MockChainRead, Error P.Ledger.ToCardanoError] effs) =>
   TxSkel ->
@@ -24,9 +24,9 @@ toInsReference skel = do
   -- As regular inputs can be used to hold scripts as if in reference inputs, we
   -- need to remove from the reference inputs stored in redeemers the ones that
   -- already appear in the inputs to avoid validation errors.
-  let indirectReferenceInputs = txSkelInsReferenceInRedeemers skel
-      redundantReferenceInputs = indirectReferenceInputs `Set.intersection` Map.keysSet (txSkelIns skel)
-      refInputs = Set.toList (txSkelInsReference skel <> indirectReferenceInputs `Set.difference` redundantReferenceInputs)
+  let indirectReferenceInputs = txSkelReferenceInputsInRedeemers skel
+      redundantReferenceInputs = indirectReferenceInputs `Set.intersection` Map.keysSet (txSkelInputs skel)
+      refInputs = Set.toList (txSkelReferenceInputs skel <> indirectReferenceInputs `Set.difference` redundantReferenceInputs)
   if null refInputs
     then return Cardano.TxInsReferenceNone
     else do