+++ src/Cooked/Skeleton.hs
@@ -14,24 +14,34 @@
 -- that are less relevant to testing, such as collaterals or fees
 module Cooked.Skeleton
   ( module X,
+
+    -- * Data type
     TxSkel (..),
+
+    -- * Optics
     txSkelLabelsL,
     txSkelOptsL,
     txSkelMintsL,
     txSkelValidityRangeL,
     txSkelProposalsL,
     txSkelSignatoriesL,
-    txSkelInsL,
-    txSkelInsReferenceL,
-    txSkelOutsL,
+    txSkelInputsL,
+    txSkelReferenceInputsL,
+    txSkelOutputsL,
     txSkelWithdrawalsL,
     txSkelCertificatesL,
+    txSkelRedeemersT,
+    txSkelRedeemedScriptsT,
+
+    -- * Smart constructor
     txSkelTemplate,
+
+    -- * Utilities
     txSkelKnownTxOutRefs,
     txSkelWithdrawnValue,
     txSkelWithdrawingScripts,
-    txSkelValueInOutputs,
-    txSkelInsReferenceInRedeemers,
+    txSkelPaidValue,
+    txSkelReferenceInputsInRedeemers,
     txSkelProposingScripts,
     txSkelMintingScripts,
     txSkelCertifyingScripts,
@@ -91,16 +101,16 @@ data TxSkel where
       --
       -- - On 'Api.TxOutRef's referencing UTxOs belonging to scripts, use
       --   the 'Cooked.Skeleton.Redeemer.someTxSkelRedeemer' smart constructor.
-      txSkelIns :: Map Api.TxOutRef TxSkelRedeemer,
+      txSkelInputs :: Map Api.TxOutRef TxSkelRedeemer,
       -- | All outputs directly referenced by the transaction. Each of them will
       -- be directly translated into a Cardano reference input. Additional
       -- reference inputs can be found within the various redeemers of the
       -- skeleton to host reference scripts. Function
-      -- 'txSkelInsReferenceInRedeemers' collects those all.
-      txSkelInsReference :: Set Api.TxOutRef,
+      -- 'txSkelReferenceInputsInRedeemers' collects those all.
+      txSkelReferenceInputs :: Set Api.TxOutRef,
       -- | The outputs of the transaction. These will occur in exactly this
       -- order on the transaction.
-      txSkelOuts :: [TxSkelOut],
+      txSkelOutputs :: [TxSkelOut],
       -- | Possible proposals issued in this transaction to be voted on and
       -- possible enacted later on.
       txSkelProposals :: [TxSkelProposal],
@@ -124,26 +134,47 @@ makeLensesFor [("txSkelMints", "txSkelMintsL")] ''TxSkel
 -- | Focuses on the validity range of a 'TxSkel'
 makeLensesFor [("txSkelValidityRange", "txSkelValidityRangeL")] ''TxSkel
 
+-- | Focuses on the proposals of a 'TxSkel'
+makeLensesFor [("txSkelProposals", "txSkelProposalsL")] ''TxSkel
+
 -- | Focuses on the signatories of a 'TxSkel'
 makeLensesFor [("txSkelSignatories", "txSkelSignatoriesL")] ''TxSkel
 
 -- | Focuses on the inputs of a 'TxSkel'
-makeLensesFor [("txSkelIns", "txSkelInsL")] ''TxSkel
+makeLensesFor [("txSkelInputs", "txSkelInputsL")] ''TxSkel
 
 -- | Focuses on the reference inputs of a 'TxSkel'
-makeLensesFor [("txSkelInsReference", "txSkelInsReferenceL")] ''TxSkel
+makeLensesFor [("txSkelReferenceInputs", "txSkelReferenceInputsL")] ''TxSkel
 
 -- | Focuses on the outputs of a 'TxSkel'
-makeLensesFor [("txSkelOuts", "txSkelOutsL")] ''TxSkel
-
--- | Focuses on the proposals of a 'TxSkel'
-makeLensesFor [("txSkelProposals", "txSkelProposalsL")] ''TxSkel
+makeLensesFor [("txSkelOutputs", "txSkelOutputsL")] ''TxSkel
 
 -- | Focuses on the withdrawals of a 'TxSkel'
 makeLensesFor [("txSkelWithdrawals", "txSkelWithdrawalsL")] ''TxSkel
 
 -- | Focuses on the certificates of a 'TxSkel'
 makeLensesFor [("txSkelCertificates", "txSkelCertificatesL")] ''TxSkel
+
+-- | A traversal focusing every 'TxSkelRedeemer' of a 'TxSkel', in all five
+-- positions (spending, minting, proposing, withdrawing and certifying).
+txSkelRedeemersT :: Traversal' TxSkel TxSkelRedeemer
+txSkelRedeemersT =
+  (txSkelInputsL % iso Map.toList Map.fromList % traversed % _2)
+    `adjoin` (txSkelMintsL % txSkelMintsListI % traversed % mintRedeemedScriptL % userTxSkelRedeemerL)
+    `adjoin` (txSkelProposalsL % traversed % txSkelProposalMConstitutionAT % _Just % userTxSkelRedeemerL)
+    `adjoin` (txSkelWithdrawalsL % txSkelWithdrawalsListI % traversed % withdrawalUserL % userTxSkelRedeemerAT)
+    `adjoin` (txSkelCertificatesL % traversed % txSkelCertificateOwnerAT @IsEither % userTxSkelRedeemerAT)
+
+-- | A traversal focusing every script redeemed directly within a 'TxSkel',
+-- that is in the minting, proposing, withdrawing and certifying positions. The
+-- spending position is excluded, as the scripts spent there are not stored in
+-- the skeleton but fetched from the index based on the inputs' references.
+txSkelRedeemedScriptsT :: Traversal' TxSkel (User IsScript Redemption)
+txSkelRedeemedScriptsT =
+  (txSkelMintsL % txSkelMintsListI % traversed % mintRedeemedScriptL)
+    `adjoin` (txSkelProposalsL % traversed % txSkelProposalConstitutionAT)
+    `adjoin` (txSkelWithdrawalsL % txSkelWithdrawalsListI % traversed % withdrawalUserL % userEitherScriptP)
+    `adjoin` (txSkelCertificatesL % traversed % txSkelCertificateOwnerAT @IsEither % userEitherScriptP)
 
 -- | A convenience template of an empty transaction skeleton.
 txSkelTemplate :: TxSkel
@@ -154,27 +185,13 @@ txSkelTemplate =
       txSkelMints = mempty,
       txSkelValidityRange = Api.always,
       txSkelSignatories = mempty,
-      txSkelIns = mempty,
-      txSkelInsReference = mempty,
-      txSkelOuts = mempty,
+      txSkelInputs = mempty,
+      txSkelReferenceInputs = mempty,
+      txSkelOutputs = mempty,
       txSkelProposals = mempty,
       txSkelWithdrawals = mempty,
       txSkelCertificates = mempty
     }
-
--- | Returns the full value contained in the skeleton outputs
-txSkelValueInOutputs :: TxSkel -> Api.Value
-txSkelValueInOutputs = foldOf (txSkelOutsL % folded % txSkelOutValueL)
-
--- | All 'Api.TxOutRef's in reference inputs from redeemers
-txSkelInsReferenceInRedeemers :: TxSkel -> Set Api.TxOutRef
-txSkelInsReferenceInRedeemers TxSkel {..} =
-  Set.fromList $
-    toListOf (to Map.elems % traversed % txSkelRedeemerReferenceInputAT) txSkelIns
-      <> toListOf (traversed % txSkelProposalMConstitutionAT % _Just % userTxSkelRedeemerL % txSkelRedeemerReferenceInputAT) txSkelProposals
-      <> toListOf (txSkelMintsListI % traversed % mintRedeemedScriptL % userTxSkelRedeemerL % txSkelRedeemerReferenceInputAT) txSkelMints
-      <> toListOf (txSkelWithdrawalsListI % traversed % withdrawalUserL % userTxSkelRedeemerAT % txSkelRedeemerReferenceInputAT) txSkelWithdrawals
-      <> toListOf (traversed % txSkelCertificateOwnerAT % userTxSkelRedeemerL % txSkelRedeemerReferenceInputAT) txSkelCertificates
 
 -- | All `Api.TxOutRef`s known by a given transaction skeleton. This includes
 -- TxOutRef`s used as inputs of the skeleton and 'Api.TxOutRef's used as reference
@@ -182,7 +199,7 @@ txSkelInsReferenceInRedeemers TxSkel {..} =
 -- 'Api.TxOutRef's used for balancing and additional 'Api.TxOutRef's used as collateral
 -- inputs, as they are not part of the skeleton.
 txSkelKnownTxOutRefs :: TxSkel -> Set Api.TxOutRef
-txSkelKnownTxOutRefs skel@TxSkel {..} = txSkelInsReferenceInRedeemers skel <> Map.keysSet txSkelIns <> txSkelInsReference
+txSkelKnownTxOutRefs skel@TxSkel {..} = txSkelReferenceInputsInRedeemers skel <> Map.keysSet txSkelInputs <> txSkelReferenceInputs
 
 -- | Returns the total value withdrawn in this 'TxSkel'
 txSkelWithdrawnValue :: TxSkel -> Api.Value
@@ -191,6 +208,15 @@ txSkelWithdrawnValue = Script.toValue . txSkelWithdrawals
 -- | Returns all the scripts involved in withdrawals in this 'TxSkel'
 txSkelWithdrawingScripts :: TxSkel -> [VScript]
 txSkelWithdrawingScripts = toListOf (txSkelWithdrawalsL % txSkelWithdrawalsListI % traversed % withdrawalUserL % userVScriptAT)
+
+-- | Returns the full value contained in the skeleton outputs
+txSkelPaidValue :: TxSkel -> Api.Value
+txSkelPaidValue = foldOf (txSkelOutputsL % folded % txSkelOutValueL)
+
+-- | All 'Api.TxOutRef's in reference inputs from redeemers
+txSkelReferenceInputsInRedeemers :: TxSkel -> Set Api.TxOutRef
+txSkelReferenceInputsInRedeemers =
+  Set.fromList . toListOf (txSkelRedeemersT % txSkelRedeemerReferenceInputAT)
 
 -- | Returns all the scripts involved in proposals in this 'TxSkel'
 txSkelProposingScripts :: TxSkel -> [VScript]