+++ src/Cooked/Tweak/Inputs.hs
@@ -3,8 +3,7 @@
 module Cooked.Tweak.Inputs
   ( ensureInputTweak,
     addInputTweak,
-    removeInputTweak,
-    modifySpendRedeemersOfTypeTweak,
+    removeInputsTweak,
   )
 where
 
@@ -12,7 +11,6 @@ import Control.Monad
 import Cooked.Skeleton
 import Cooked.Tweak.Common
 import Data.Map qualified as Map
-import Optics.Core
 import PlutusLedgerApi.V3 qualified as Api
 import Polysemy
 import Polysemy.NonDet
@@ -26,11 +24,11 @@ ensureInputTweak ::
   TxSkelRedeemer ->
   Sem effs (Maybe (Api.TxOutRef, TxSkelRedeemer))
 ensureInputTweak oref howConsumed = do
-  presentInputs <- viewTweak txSkelInsL
+  presentInputs <- viewTweak txSkelInputsL
   if presentInputs Map.!? oref == Just howConsumed
     then return Nothing
     else do
-      overTweak txSkelInsL (Map.insert oref howConsumed)
+      overTweak txSkelInputsL (Map.insert oref howConsumed)
       return $ Just (oref, howConsumed)
 
 -- | Add an input to a transaction. If the given 'Api.TxOutRef' is already being
@@ -41,34 +39,18 @@ addInputTweak ::
   TxSkelRedeemer ->
   Sem effs ()
 addInputTweak oref howConsumed = do
-  presentInputs <- viewTweak txSkelInsL
+  presentInputs <- viewTweak txSkelInputsL
   guard (Map.notMember oref presentInputs)
-  overTweak txSkelInsL (Map.insert oref howConsumed)
+  overTweak txSkelInputsL (Map.insert oref howConsumed)
 
 -- | Remove transaction inputs according to a given predicate. The returned list
 -- contains all removed inputs.
-removeInputTweak ::
+removeInputsTweak ::
   (Member Tweak effs) =>
   (Api.TxOutRef -> TxSkelRedeemer -> Bool) ->
   Sem effs [(Api.TxOutRef, TxSkelRedeemer)]
-removeInputTweak removePred = do
-  presentInputs <- viewTweak txSkelInsL
+removeInputsTweak removePred = do
+  presentInputs <- viewTweak txSkelInputsL
   let (removed, kept) = Map.partitionWithKey removePred presentInputs
-  setTweak txSkelInsL kept
+  setTweak txSkelInputsL kept
   return $ Map.toList removed
-
--- | Applies an optional modification to all spend redeemers of type a. Returns
--- the list of modified spending redemeers, as they were before being modified.
-modifySpendRedeemersOfTypeTweak ::
-  forall a b effs.
-  ( RedeemerConstrs a,
-    RedeemerConstrs b,
-    Member Tweak effs
-  ) =>
-  (a -> Maybe b) ->
-  Sem effs [TxSkelRedeemer]
-modifySpendRedeemersOfTypeTweak f =
-  overMaybeTweak (txSkelInsL % iso Map.toList Map.fromList % traversed % _2) $ \red -> do
-    typedRedeemer <- preview txSkelRedeemerTypedAT red
-    typedRedeemerModified <- f typedRedeemer
-    return $ set (txSkelRedeemerTypedAT @a) typedRedeemerModified red